+++ richat/src/grpc/server.rs
@@ -75,7 +75,7 @@ pub struct GrpcServer {
     messages: Messages,
     block_meta: Option<Arc<BlockMetaStorage>>,
     filter_limits: Arc<ConfigFilterLimits>,
-    ping_iterval: Duration,
+    ping_interval: Duration,
     subscribe_id: Arc<AtomicU64>,
     subscribe_clients: Arc<Mutex<VecDeque<SubscribeClient>>>,
     subscribe_messages_len_max: usize,
@@ -120,7 +120,7 @@ impl GrpcServer {
             messages,
             block_meta,
             filter_limits: Arc::new(config.filter_limits),
-            ping_iterval: config.stream.ping_iterval,
+            ping_interval: config.stream.ping_interval,
             subscribe_id: Arc::new(AtomicU64::new(0)),
             subscribe_clients: Arc::new(Mutex::new(VecDeque::new())),
             subscribe_messages_len_max: config.stream.messages_len_max,
@@ -161,11 +161,13 @@ impl GrpcServer {
         let server = tokio::spawn(async move {
             if let Err(error) = server_builder
                 .layer(InterceptorLayer::new(move |request: Request<()>| {
-                    if config.x_token.is_empty() {
+                    if config.x_tokens.is_empty() {
                         Ok(request)
                     } else {
                         match request.metadata().get("x-token") {
-                            Some(token) if config.x_token.contains(token.as_bytes()) => Ok(request),
+                            Some(token) if config.x_tokens.contains(token.as_bytes()) => {
+                                Ok(request)
+                            }
                             _ => Err(Status::unauthenticated("No valid auth token")),
                         }
                     }
@@ -416,7 +418,7 @@ impl GrpcServer {
 
         tokio::spawn({
             let shutdown = self.shutdown.clone();
-            let ping_interval = self.ping_iterval;
+            let ping_interval = self.ping_interval;
             let client = client.clone();
             async move {
                 let mut ts_latest = Instant::now();