+++ CHANGELOG.md
@@ -26,6 +26,11 @@ The minor version will be incremented upon a breaking change and the patch versi
 - richat-v6.0.0
 - richat-shared-v6.0.0
 
+### Fixes
+
+- richat: fix `ping_interval` typo ([#148](https://github.com/lamports-dev/richat/pull/148))
+- richat: fix `x_tokens` typo ([#149](https://github.com/lamports-dev/richat/pull/149))
+
 ### Breaking
 
 - richat: upgrade to agave 3.0 ([#146](https://github.com/lamports-dev/richat/pull/146))